chore: move printer impl into multiple files
